requirements.txt
asgiref==3.8.1
Django==5.0.6
djangorestframework==3.15.2
djangorestframework-simplejwt==5.2.2
django-phonenumber-field==8.0.0
greenlet==3.0.3
gunicorn==22.0.0
line-bot-sdk==3.12.0
matplotlib==3.9.1
numpy==1.26.4
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
pandas==2.2.2
phonenumbers==8.13.46
psycopg2-binary==2.9.6
pyee==11.1.0
python-dotenv==1.0.0
PyYAML==6.0.1
requests==2.32.3
scikit-image==0.24.0
scipy==1.14.0
seaborn==0.13.2
sqlparse==0.5.0
tqdm==4.66.4
typing_extensions==4.12.0
tzdata==2024.1
